test/test_sam/requirements.txt
torch>=2.3.1
torchvision>=0.18.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
matplotlib>=3.9.1
jupyter>=1.0.0
opencv-python>=4.7.0
black==24.2.0
usort==1.0.2
ufmt==2.0.0b2
pandas>=1.1.4
seaborn>=0.11.0
gradio>=3.35.2
ultralytics>=8.0.120
flake8
isort
onnx
onnxruntime-gpu
pycocotools
